Filter on assigned, priority and status
notes:

Line 4: We change the helper path to be absolute to make testing easier
Line 6 (removed): The cpf namespace isn't used
Lines 30-32: We add default values so that we're more flexible to what we're passed
    and don't have to update the editor/public ui in sync
Lines 50-52: Nothing interesting
Lines 84-85: We only apply this test if we're already allowing unpublished stuff,
    because we don't want to leak this information to the public.
Line 120: add the new queries to the query stack

Lines 87-117:
We build three queries --
* is there not an assigned-to at all? If so, it's new.
  (New could also be "has a blank assigned-to", but that makes things more complicated,
  so we'll need to delete the assigned-to instead. For some reason assigning-to No-one
  fails at the moment, need to investigate this.)
* Is there an assigned-to and it's held? It's held.
* Is it assigned to and it isn't held? It's in progress.
  (We use "not (held-is-true)" to cover the case where editor-hold doesn't exist)
Then we pick one based on the `editor-status` value, or don't if:
* there is no editor-status value
* we're not allowing unpublished
* we don't understand the editor-status value.

Profitable bikeshedding:
Are `held`/`inprogress`/`new` the best terms?
Should we factor the complicated function out into a helper?
Should I do something about 'assigned-to' potentially being the empty string? (And the comment)